requirements.txt
alabaster==0.7.9
Babel==2.3.4
backports.pbkdf2==0.1
click==6.6
coverage==4.1
coveralls==1.1
docopt==0.6.2
docutils==0.13.1
Flask==0.12
Flask-Login==0.3.2
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-Testing==0.5.0
Foundations==2.1.0
imagesize==0.7.1
itsdangerous==0.24
Jinja2==2.8
linecache2==1.0.0
Manager==2.0.5
MarkupSafe==0.23
Oncilla==0.1.0
Pygments==2.1.3
pytz==2016.10
requests==2.10.0
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.5.1
sphinx-rtd-theme==0.1.9
SQLAlchemy==1.0.14
traceback2==1.4.0
unittest2==1.1.0
Werkzeug==0.11.11
